Refactor EditUserDialog to be it's own component.
